refactor: split SdkLoader into SdkLoader, SdkBootstrap, SdkDiscovery
